requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.5
django-redis==5.0.0
django-redis-cache==3.0.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
environs==9.3.2
idna==3.2
kombu==5.1.0
marshmallow==3.13.0
Pillow==8.3.1
prompt-toolkit==3.0.19
pycrypto==2.6.1
PyJWT==2.1.0
python-dotenv==0.19.0
python-redis-lock==3.7.0
pytz==2021.1
redis==3.5.3
redis-cache==0.1.5
redis-simple-cache==0.0.8
redis-structures==0.1.7
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
vine==5.0.0
vital-tools==0.1.13
wcwidth==0.2.5

django-filter~=2.4.0
django-mptt~=0.13.3